+++ website/templates/base.html
@@ -35,15 +35,15 @@
             <div id="sidebarContent">
                 <div id="sidebarSubcontent">
                     <h2>Next Feed Time:</h2>
-                    <p>TIME GOES HERE</p>
+                    <p>{{ nextFeedTime }}</p>
                 </div>
                 <div id="sidebarSubcontent">
                     <h2>Last Feed Time:</h2>
-                    <p>TIME GOES HERE</p>
+                    <p>{{ lastFeedTime }}</p>
                 </div>
                 <div id="sidebarSubcontent">
                     <h2>Feeder Status:</h2>
-                    <p>STATUS GOES HERE</p>
+                    <p>{{ status }}</p>
                 </div>
             </div>
         </aside>
